refactor: 검증 함수의 export 방식을 기본으로 변경
